Remove deprecation warnings in 1.37
From 1.37, the template directory should be relative
to the skin not core and will begin omitting deprecation warnings.

Note this change will break compatibility with 1.36 so please
exercise caution when deploying this :)

If you need to maintain compatibility with 1.36, you can set this
inside SkinWMAU constructor safely.

See T262067 for more information

Change-Id: I09cf50345b289779096d3080cb881022268ca23e