Allow 'icon' menu config to refer to a wiki file
Check the menu items' 'icon' attribute to see if there's a file
with that name, and if found use it instead of checking for a
feathericon in the skin's resources.

Cache the icon's thumbnail URL for a month.

Adds a new 'icon_width' parameter for menu items, defaulting
to 24 px.

Bug: T285172
Change-Id: Ic7ac1a6119ca121470783f60fccd9d4dceaf71d9